[FIX] PHPStan level 5 fixes and EventController regression fix
- Cast date() string returns to int for mktime() calls in TCA range
  configs (11 files)
- Simplify EventDemandFactory/PerformanceDemandFactory: remove
  always-true instanceof checks on PeriodAwareDemandInterface
- Simplify DemandedRepositoryTrait::__call(): remove always-true
  class_parents() !== false check
- Remove stale @phpstan-ignore-next-line comments in
  NotificationService::duplicate() for errors that no longer occur
- Migrate MetaTagViewHelper to Fluid v4 attribute API
  (getAttribute(ServerRequestInterface::class) instead of the
  removed RenderingContextInterface::getRequest())
- Fix EventController::showAction(): restore the null-check branch
  that was accidentally collapsed by an earlier Rector pass, which
  made every event detail page redirect to the list view instead of
  rendering the detail template

+++ Configuration/TCA/tx_t3events_domain_model_ticketclass.php
@@ -91,7 +91,7 @@ return [
                 'checkbox' => 0,
                 'default' => 0,
                 'range' => [
-                    'lower' => mktime(0, 0, 0, date('m'), date('d'), date('Y'))
+                    'lower' => mktime(0, 0, 0, (int)date('m'), (int)date('d'), (int)date('Y'))
                 ],
                 'behaviour' => [
                     'allowLanguageSynchronization' => true
@@ -107,7 +107,7 @@ return [
                 'checkbox' => 0,
                 'default' => 0,
                 'range' => [
-                    'lower' => mktime(0, 0, 0, date('m'), date('d'), date('Y'))
+                    'lower' => mktime(0, 0, 0, (int)date('m'), (int)date('d'), (int)date('Y'))
                 ],
                 'behaviour' => [
                     'allowLanguageSynchronization' => true
